archive: use files instead of ipld
