+++ archive/archive.go
@@ -2,16 +2,16 @@
 package archive
 
 import (
+	"archive/tar"
 	"bufio"
 	"compress/gzip"
-	"context"
+	"errors"
+	"fmt"
 	"io"
 	"path"
+	"time"
 
-	tar "github.com/ipfs/go-unixfs/archive/tar"
-	uio "github.com/ipfs/go-unixfs/io"
-
-	ipld "github.com/ipfs/go-ipld-format"
+	files "github.com/ipfs/go-ipfs-files"
 )
 
 // DefaultBufSize is the buffer size for gets. for now, 1MB, which is ~4 blocks.
@@ -31,8 +31,7 @@ func (i *identityWriteCloser) Close() error {
 }
 
 // DagArchive is equivalent to `ipfs getdag $hash | maybe_tar | maybe_gzip`
-func DagArchive(ctx context.Context, nd ipld.Node, name string, dag ipld.DAGService, archive bool, compression int) (io.Reader, error) {
-
+func FileArchive(f files.Node, name string, archive bool, compression int) (io.Reader, error) {
 	cleaned := path.Clean(name)
 	_, filename := path.Split(cleaned)
 
@@ -67,13 +66,13 @@ func DagArchive(ctx context.Context, nd ipld.Node, name string, dag ipld.DAGServ
 
 	if !archive && compression != gzip.NoCompression {
 		// the case when the node is a file
-		dagr, err := uio.NewDagReader(ctx, nd, dag)
-		if checkErrAndClosePipe(err) {
-			return nil, err
+		r := files.ToFile(f)
+		if r == nil {
+			return nil, errors.New("file is not regular")
 		}
 
 		go func() {
-			if _, err := dagr.WriteTo(maybeGzw); checkErrAndClosePipe(err) {
+			if _, err := io.Copy(maybeGzw, r); checkErrAndClosePipe(err) {
 				return
 			}
 			closeGzwAndPipe() // everything seems to be ok
@@ -82,14 +81,14 @@ func DagArchive(ctx context.Context, nd ipld.Node, name string, dag ipld.DAGServ
 		// the case for 1. archive, and 2. not archived and not compressed, in which tar is used anyway as a transport format
 
 		// construct the tar writer
-		w, err := tar.NewWriter(ctx, dag, maybeGzw)
+		w, err := NewWriter(maybeGzw)
 		if checkErrAndClosePipe(err) {
 			return nil, err
 		}
 
 		go func() {
 			// write all the nodes recursively
-			if err := w.WriteNode(nd, filename); checkErrAndClosePipe(err) {
+			if err := w.WriteFile(f, filename); checkErrAndClosePipe(err) {
 				return
 			}
 			w.Close()         // close tar writer
@@ -106,3 +105,96 @@ func newMaybeGzWriter(w io.Writer, compression int) (io.WriteCloser, error) {
 	}
 	return &identityWriteCloser{w}, nil
 }
+
+type Writer struct {
+
+	TarW *tar.Writer
+}
+
+// NewWriter wraps given io.Writer.
+func NewWriter(w io.Writer) (*Writer, error) {
+	return &Writer{
+		TarW: tar.NewWriter(w),
+	}, nil
+}
+
+func (w *Writer) writeDir(f files.Directory, fpath string) error {
+	if err := writeDirHeader(w.TarW, fpath); err != nil {
+		return err
+	}
+
+	it := f.Entries()
+	for it.Next() {
+		if err := w.WriteFile(it.Node(), path.Join(fpath, it.Name())); err != nil {
+			return err
+		}
+	}
+	return it.Err()
+}
+
+func (w *Writer) writeFile(f files.File, fpath string) error {
+	size, err := f.Size()
+	if err != nil {
+		return err
+	}
+
+	if err := writeFileHeader(w.TarW, fpath, uint64(size)); err != nil {
+		return err
+	}
+
+	if _, err := io.Copy(w.TarW, f); err != nil {
+		return err
+	}
+	w.TarW.Flush()
+	return nil
+}
+
+// WriteNode adds a node to the archive.
+func (w *Writer) WriteFile(nd files.Node, fpath string) error {
+	switch nd := nd.(type) {
+	case *files.Symlink:
+		return writeSymlinkHeader(w.TarW, nd.Target, fpath)
+	case files.File:
+		return w.writeFile(nd, fpath)
+	case files.Directory:
+		return w.writeDir(nd, fpath)
+	default:
+		return fmt.Errorf("file type %T is not supported", nd)
+	}
+}
+
+// Close closes the tar writer.
+func (w *Writer) Close() error {
+	return w.TarW.Close()
+}
+
+func writeDirHeader(w *tar.Writer, fpath string) error {
+	return w.WriteHeader(&tar.Header{
+		Name:     fpath,
+		Typeflag: tar.TypeDir,
+		Mode:     0777,
+		ModTime:  time.Now(),
+		// TODO: set mode, dates, etc. when added to unixFS
+	})
+}
+
+func writeFileHeader(w *tar.Writer, fpath string, size uint64) error {
+	return w.WriteHeader(&tar.Header{
+		Name:     fpath,
+		Size:     int64(size),
+		Typeflag: tar.TypeReg,
+		Mode:     0644,
+		ModTime:  time.Now(),
+		// TODO: set mode, dates, etc. when added to unixFS
+	})
+}
+
+func writeSymlinkHeader(w *tar.Writer, target, fpath string) error {
+	return w.WriteHeader(&tar.Header{
+		Name:     fpath,
+		Linkname: target,
+		Mode:     0777,
+		Typeflag: tar.TypeSymlink,
+	})
+}
+
